Move to the right folder structure for an editable pip package and make imports relative among the different scripts
