+++ src/nllb_try/evaluate.py
@@ -2,9 +2,9 @@ import os
 import pandas as pd
 import matplotlib.pyplot as plt
 from sacrebleu import corpus_bleu, corpus_chrf
-from tokenizer_and_model_setup import setup_model_and_tokenizer, cleanup
-from train import preproc  # Import the preprocessing from train script
-from config import config
+from .tokenizer_and_model_setup import setup_model_and_tokenizer, cleanup
+from .train import preproc  # Import the preprocessing from train script
+from .config import config
 
 
 def translate(text, src_lang: str, tgt_lang: str, model, tokenizer, a=16, b=1.5, max_input_length: int = 200, **kwargs):
